+++ src/app/loadout-builder/filter/ArmorUpgradePicker.tsx
@@ -1,6 +1,7 @@
 import { UpgradeSpendTier } from '@destinyitemmanager/dim-api-types';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
+import CheckButton from 'app/dim-ui/CheckButton';
 import Sheet from 'app/dim-ui/Sheet';
 import { t } from 'app/i18next-t';
 import { useD2Definitions } from 'app/manifest/selectors';
@@ -30,6 +31,7 @@ function getDisplayProperties(
     UpgradeMaterialHashes.ascendantShard
   ).displayProperties;
   switch (upgradeSpendTier) {
+    case UpgradeSpendTier.AscendantShardsLockEnergyType:
     case UpgradeSpendTier.Nothing:
       return { name: t('LoadoutBuilder.NoUpgrades') };
     case UpgradeSpendTier.LegendaryShards: {
@@ -50,12 +52,6 @@ function getDisplayProperties(
         icon: ascendantShardDisplay.icon,
       };
     }
-    case UpgradeSpendTier.AscendantShardsLockEnergyType: {
-      return {
-        name: t('LoadoutBuilder.LockElement', { material: ascendantShardDisplay.name }),
-        icon: ascendantShardDisplay.icon,
-      };
-    }
     case UpgradeSpendTier.AscendantShards: {
       return ascendantShardDisplay;
     }
@@ -65,15 +61,18 @@ function getDisplayProperties(
 export function SelectedArmorUpgrade({
   defs,
   upgradeSpendTier,
+  lockItemEnergyType,
 }: {
   defs: D2ManifestDefinitions;
   upgradeSpendTier: UpgradeSpendTier;
+  lockItemEnergyType: boolean;
 }) {
   const { name, icon } = getDisplayProperties(defs, upgradeSpendTier);
+  const displayName = lockItemEnergyType ? `${name} + ${t('LoadoutBuilder.LockElement')}` : name;
   return (
     <div className={styles.selectedUpgradeOption}>
       {icon && <BungieImage src={icon} />}
-      <div>{name}</div>
+      <div>{displayName}</div>
     </div>
   );
 }
@@ -81,25 +80,26 @@ export function SelectedArmorUpgrade({
 /** A drawer to select an exotic for your build. */
 function ArmorUpgradePicker({
   currentUpgradeSpendTier,
+  lockItemEnergyType,
   onClose,
 }: {
   currentUpgradeSpendTier: UpgradeSpendTier;
+  lockItemEnergyType: boolean;
   onClose(): void;
 }) {
   const defs = useD2Definitions()!;
   const setSetting = useSetSetting();
 
   const upgradeOptions: { value: UpgradeSpendTier; content: ReactNode }[] = useMemo(() => {
+    const elementLockDescription = lockItemEnergyType
+      ? t('LoadoutBuilder.ElementLockedDescription')
+      : t('LoadoutBuilder.ElementNotLockedDescription');
     const legendaryShardDisplay = getDisplayProperties(defs, UpgradeSpendTier.LegendaryShards);
     const enhancementPrismDisplay = getDisplayProperties(defs, UpgradeSpendTier.EnhancementPrisms);
     const notExoticDisplay = getDisplayProperties(defs, UpgradeSpendTier.AscendantShardsNotExotic);
     const notMasterworkedDisplay = getDisplayProperties(
       defs,
       UpgradeSpendTier.AscendantShardsNotMasterworked
-    );
-    const lockedEnergyTypeDisplay = getDisplayProperties(
-      defs,
-      UpgradeSpendTier.AscendantShardsLockEnergyType
     );
     const ascendantShardDisplay = getDisplayProperties(defs, UpgradeSpendTier.AscendantShards);
 
@@ -121,6 +121,7 @@ function ArmorUpgradePicker({
             icon={legendaryShardDisplay.icon}
             details={t('LoadoutBuilder.LegendaryShardsAndEnhancementPrismDetails', {
               energyLevel: 7, // todo (ryanr) generate or obtain this programatically
+              elementLockDescription,
             })}
           />
         ),
@@ -133,6 +134,7 @@ function ArmorUpgradePicker({
             icon={enhancementPrismDisplay.icon}
             details={t('LoadoutBuilder.LegendaryShardsAndEnhancementPrismDetails', {
               energyLevel: 9,
+              elementLockDescription,
             })}
           />
         ),
@@ -143,7 +145,10 @@ function ArmorUpgradePicker({
           <UpgradeOption
             name={notExoticDisplay.name}
             icon={notExoticDisplay.icon}
-            details={t('LoadoutBuilder.AscendantShardNotExoticDetails', { energyLevel: 9 })}
+            details={t('LoadoutBuilder.AscendantShardNotExoticDetails', {
+              energyLevel: 9,
+              elementLockDescription,
+            })}
           />
         ),
       },
@@ -153,17 +158,9 @@ function ArmorUpgradePicker({
           <UpgradeOption
             name={notMasterworkedDisplay.name}
             icon={notMasterworkedDisplay.icon}
-            details={t('LoadoutBuilder.AscendantShardNotMasterworkedDetails')}
-          />
-        ),
-      },
-      {
-        value: UpgradeSpendTier.AscendantShardsLockEnergyType,
-        content: (
-          <UpgradeOption
-            name={lockedEnergyTypeDisplay.name}
-            icon={lockedEnergyTypeDisplay.icon}
-            details={t('LoadoutBuilder.AscendantShardLockEnergyDetails')}
+            details={t('LoadoutBuilder.AscendantShardNotMasterworkedDetails', {
+              elementLockDescription,
+            })}
           />
         ),
       },
@@ -173,12 +170,12 @@ function ArmorUpgradePicker({
           <UpgradeOption
             name={ascendantShardDisplay.name}
             icon={ascendantShardDisplay.icon}
-            details={t('LoadoutBuilder.AscendantShardDetails')}
+            details={t('LoadoutBuilder.AscendantShardDetails', { elementLockDescription })}
           />
         ),
       },
     ];
-  }, [defs]);
+  }, [defs, lockItemEnergyType]);
 
   return (
     <Sheet
@@ -193,6 +190,16 @@ function ArmorUpgradePicker({
     >
       {({ onClose }) => (
         <div className={styles.container}>
+          <div className={styles.modifiers}>
+            <CheckButton
+              name="lo-lock-item-energy-type"
+              className={styles.lockEnergyType}
+              checked={lockItemEnergyType}
+              onChange={(checked) => setSetting('loLockItemEnergyType', checked)}
+            >
+              {t('LoadoutBuilder.LockElement')}
+            </CheckButton>
+          </div>
           <div className={styles.items}>
             {upgradeOptions.map((option) => (
               <div