+++ src/app/loadout-builder/generated-sets/GeneratedSet.tsx
@@ -29,6 +29,7 @@ interface Props {
   params: LoadoutParameters;
   halfTierMods: PluggableInventoryItemDefinition[];
   upgradeSpendTier: UpgradeSpendTier;
+  lockItemEnergyType: boolean;
 }
 
 /**
@@ -49,6 +50,7 @@ function GeneratedSet({
   params,
   halfTierMods,
   upgradeSpendTier,
+  lockItemEnergyType,
 }: Props) {
   const defs = useD2Definitions();
   // Set the loadout property to show/hide the loadout menu
@@ -68,7 +70,8 @@ function GeneratedSet({
     defs,
     set.armor.map((items) => items[0]),
     lockedMods,
-    upgradeSpendTier
+    upgradeSpendTier,
+    lockItemEnergyType
   );
 
   const canCompareLoadouts =